+++ VisitorsTracker/appsettings.json
@@ -6,6 +6,13 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "JWTOptions": {
+    "SecretKey": "0d5b3235a8b403c3dab9c3f4f65c07fcalskd234n1k41230",
+    "LifeTime": 1800,
+    "GoogleClientId": "462310270689-sqldnkkc44afqoirqqbatjit4hl8i97h.apps.googleusercontent.com",
+    "GoogleClientSecret": "8rg6cEvMPT_pXiQFWxFGN7BM",
+    "JwtEmailEncryption": "YOUR JWT EMAIL ENCRYPTION"
+  },
   "AllowedHosts": "*",
   "ConnectionStrings": {
     "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=VisitorsTracker;Trusted_Connection=True;MultipleActiveResultSets=true"
