--- VisitorsTracker/libman.json
@@ -1,5 +1,0 @@
-{
-  "version": "1.0",
-  "defaultProvider": "cdnjs",
-  "libraries": []
-}
